// PropertyRenting.UI/ClientApp/src/app/Models/Reports/building-transaction.ts
export interface BuildingTransaction {
    type: number;
    debitAmount?: number;
    creditAmount?: number;
    autoNumber?: number;
    voucherId?: string;
    voucherDate?: Date;
    referenceId?: any;
    referenceType?: string;
    referenceAutoNumber?: number;
    referenceManualNumber?: string;
    description?: string;
}