// PropertyRenting.UI/ClientApp/src/app/Models/Reports/renter-due-installments.ts
export interface RenterDueInstallments {
    autoNumber: number;
    contractNumber: string;
    typeAR: string;
    typeEN: string;
    amount: number;
    paidAmount: number;
    balance: number;
    dueDate: Date;
    renterAR: string;
    renterEN: string;
    buildingName: string;
    unitName: string;
    unitNumber: string;
}