// PropertyRenting.UI/ClientApp/src/app/Models/Reports/unit-transaction.ts
export interface UnitTransaction {
    type: number;
    debitAmount?: number;
    creditAmount?: number;
    autoNumber?: number;
    voucherId?: string;
    voucherDate?: Date;
    referenceId?: any;
    referenceType?: string;
    referenceAutoNumber?: number;
    referenceManualNumber?: string;
    description?: string;
    buildingName?: string;
}